+++ RemedyFriends/Assets/Scripts/Auth.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ce101013bee8cda4d8426505c877dc9b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
