+++ RemedyFriends/Assets/Scripts/Auth/PopupChange.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73ccd2decd7c8e84da5a206c73bbb761
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
