+++ RemedyFriends/Assets/Scripts/Auth/Select_char.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96325182f1193de4dade7b0f7fdf7267
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
