+++ RemedyFriends/Assets/Scripts/Auth/Sign_up.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28b7e097558e9914bb500aa596ace94f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
