+++ RemedyFriends/Assets/Scripts/Auth/Users_login.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 81210c64712f7044197cd250659a7443
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
