--- RemedyFriends/Assets/Scripts/Login.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 08f4be31b760a7a43b2ce716fc9f2cb0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
