--- RemedyFriends/Assets/Scripts/Login/Login_Demo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3abcaa2338685a0419eada30cabc3718
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
