--- RemedyFriends/Assets/Scripts/Login/PopupChange.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cf2e7f75b16b0164aa0ef7295be415ff
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
