--- RemedyFriends/Assets/Scripts/Login/Select_char.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2664930f22cf45442907c4c6f30e060e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
