+++ RemedyFriends/Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/game/game2_Hide and seek with Luna.unity
+- path: Assets/Scenes/members/members_login.unity
   isLoaded: 1
   isActive: 1
