+++ RemedyFriends/Library/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,4 +1,2 @@
 Base path: C:/Program Files/Unity/Hub/Editor/2018.4.9f1/Editor/Data
 Cmd: initializeCompiler
-Cmd: initializeCompiler
-Cmd: initializeCompiler
